+++ src/extension.ts
@@ -6,6 +6,9 @@ import { start, exit } from "./domain";
 
 export function activate(context: vscode.ExtensionContext) {
   const repository = new VSCodeRepository(context);
+  var curSlideUri: vscode.Uri;
+  var curSlideLanguageId: string;
+  var useMdPreviewSetting: boolean;
 
   const toggleSlides = vscode.commands.registerCommand(
     "slides.toggle",
@@ -17,6 +20,18 @@ export function activate(context: vscode.ExtensionContext) {
         await exit(vscodeEditor, repository);
       } else {
         await start(vscodeEditor, repository);
+        useMdPreviewSetting = vscode.workspace.getConfiguration("slides")
+          .useMdPreview;
+        let activeWindow = vscode.window.activeTextEditor;
+
+        if (useMdPreviewSetting && activeWindow !== undefined) {
+          curSlideUri = activeWindow.document.uri;
+          curSlideLanguageId = activeWindow.document.languageId;
+
+          if (activeWindow.document.languageId === "markdown") {
+            await vscode.commands.executeCommand("markdown.showPreview");
+          }
+        }
       }
     }
   );
@@ -27,7 +42,22 @@ export function activate(context: vscode.ExtensionContext) {
       const { isActive } = await repository.get();
 
       if (isActive) {
+        if (useMdPreviewSetting && curSlideLanguageId === "markdown") {
+          await vscode.commands.executeCommand(
+            "workbench.action.closeActiveEditor"
+          );
+          await vscode.workspace.openTextDocument(curSlideUri);
+        }
+
         await vscode.commands.executeCommand("workbench.action.previousEditor");
+        let activeWindow = vscode.window.activeTextEditor;
+
+        if (useMdPreviewSetting && activeWindow !== undefined) {
+          curSlideLanguageId = activeWindow.document.languageId;
+          if (activeWindow.document.languageId === "markdown") {
+            await vscode.commands.executeCommand("markdown.showPreview");
+          }
+        }
       }
     }
   );
@@ -36,7 +66,22 @@ export function activate(context: vscode.ExtensionContext) {
     const { isActive } = await repository.get();
 
     if (isActive) {
+      if (useMdPreviewSetting && curSlideLanguageId === "markdown") {
+        await vscode.commands.executeCommand(
+          "workbench.action.closeActiveEditor"
+        );
+        await vscode.workspace.openTextDocument(curSlideUri);
+      }
+
       await vscode.commands.executeCommand("workbench.action.nextEditor");
+      let activeWindow = vscode.window.activeTextEditor;
+
+      if (useMdPreviewSetting && activeWindow !== undefined) {
+        curSlideLanguageId = activeWindow.document.languageId;
+        if (activeWindow.document.languageId === "markdown") {
+          await vscode.commands.executeCommand("markdown.showPreview");
+        }
+      }
     }
   });