+++ src/settings.ts
@@ -43,5 +43,8 @@ export const settings = {
   "editor.formatOnSave": false,
 
   "//k": "Disable preview so we can open all files at once",
-  "workbench.editor.enablePreview": false
+  "workbench.editor.enablePreview": false,
+
+  "//l": "Use the Markdown Preview pane to view Markdown files",
+  "workbench.useMdPreview": false
 };